ohbench/requirements.txt
numpy<2.0
scipy
pandas
tqdm
easydict
opencv-python
Pillow
moviepy
librosa
soundfile
pyloudnorm
requests
onnxruntime
decord
pyiqa
transformers>=0.34.0,<5
huggingface_hub>=0.34.0,<1.0
pytorchvideo
omegaconf
einops
blobfile
openai>=1.0
git+https://github.com/openai/CLIP.git
insightface
funasr
